Fix TPU decoding dtype issue

+++ axlearn/common/flash_attention/tpu_decoding.py
@@ -80,7 +80,7 @@ def _tpu_decoding_kernel(
     @pl.when(non_empty_kv_block_index < num_non_empty_kv_blocks)
     def compute():
         q = q_ref[...]
-        k = k_ref[...]
+        k = k_ref[...].astype(q.dtype)
         qk = pl.dot(q, k, precision=precision)
         if softmax_scale != 1.0:
             qk *= softmax_scale
@@ -109,7 +109,7 @@ def _tpu_decoding_kernel(
         l_curr = s_curr.sum(axis=-1, keepdims=True)
         l_next = l_prev_corr + l_curr
         o_prev_corr = correction * o_prev
-        v = v_ref[...]
+        v = v_ref[...].astype(q.dtype)
         o_curr = pl.dot(s_curr.astype(v.dtype), v.T, precision=precision)
 
         o_next = o_prev_corr + o_curr
@@ -185,8 +185,8 @@ def tpu_decoding(
     q = q.squeeze(1)
     # Convert to bnhs which is the native shape of KV in the kv cache. These two transposes should
     # be elided by the compiler. See `BaseQKVLinear.init_states` from attention.py.
-    k = jnp.einsum("bsnh->bnhs", k).astype(q.dtype)
-    v = jnp.einsum("bsnh->bnhs", v).astype(q.dtype)
+    k = jnp.einsum("bsnh->bnhs", k)
+    v = jnp.einsum("bsnh->bnhs", v)
     bs, kv_heads, head_dim, padded_kv_seq_len = k.shape
     if kv_seq_len is not None:
         kv_seq_len = jnp.broadcast_to(jnp.asarray(kv_seq_len), (bs,))
